working on a new restaurant adding scheme
This scheme uses text based config files for each location. These files
contain all relevant info, like name and group, and also provide a nice
way to list hours. I'm trying to make the config files handle all the
combinations of weird ways a business may be open or closed in a manner
that is easy for humans to write.